+++ modules/core/src/lib/attribute/attribute-manager.js
@@ -296,15 +296,18 @@ export default class AttributeManager {
   /* eslint-enable max-statements */
 
   _createAttribute(name, attribute, extraProps) {
+    // For expected default values see:
+    // https://github.com/visgl/luma.gl/blob/1affe21352e289eeaccee2a876865138858a765c/modules/webgl/src/classes/accessor.js#L5-L13
+    // and https://deck.gl/docs/api-reference/core/attribute-manager#add
     const props = {
       ...attribute,
       id: name,
+      isIndexed: attribute.isIndexed || attribute.elements || false,
       // Luma fields
       constant: attribute.constant || false,
-      isIndexed: attribute.isIndexed || attribute.elements,
-      size: (attribute.elements && 1) || attribute.size,
+      size: (attribute.elements && 1) || attribute.size || 1,
       value: attribute.value || null,
-      divisor: attribute.instanced || extraProps.instanced ? 1 : attribute.divisor
+      divisor: attribute.instanced || extraProps.instanced ? 1 : attribute.divisor || 0
     };
 
     return new Attribute(this.gl, props);
@@ -345,6 +348,13 @@ export default class AttributeManager {
   _updateAttribute(opts) {
     const {attribute, numInstances} = opts;
     debug(TRACE_ATTRIBUTE_UPDATE_START, attribute);
+
+    if (attribute.constant) {
+      // The attribute is flagged as constant outside of an update cycle
+      // Skip allocation and updater call
+      attribute.setConstantValue(attribute.value);
+      return;
+    }
 
     if (attribute.allocate(numInstances)) {
       debug(TRACE_ATTRIBUTE_ALLOCATE, attribute, numInstances);