+++ modules/core/src/passes/layers-pass.js
@@ -86,7 +86,7 @@ export default class LayersPass extends Pass {
           pass,
           moduleParameters
         );
-        layerParam.layerParameters = this.getLayerParameters(layer, layerIndex);
+        layerParam.layerParameters = this.getLayerParameters(layer, layerIndex, viewport);
       }
       drawLayerParams[layerIndex] = layerParam;
     }
@@ -186,7 +186,7 @@ export default class LayersPass extends Pass {
 
     let parent = layer.parent;
     while (parent) {
-      if (!parent.filterSubLayer(drawContext)) {
+      if (!parent.props.visible || !parent.filterSubLayer(drawContext)) {
         return false;
       }
       drawContext.layer = parent;