+++ modules/core/src/shaderlib/project/project.glsl.js
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import {COORDINATE_SYSTEM, PROJECTION_MODE} from '../../lib/constants';
+import {COORDINATE_SYSTEM, PROJECTION_MODE, UNIT} from '../../lib/constants';
 
 // We are generating these from the js code in constants.js
 const COORDINATE_SYSTEM_GLSL_CONSTANTS = Object.keys(COORDINATE_SYSTEM)
@@ -27,10 +27,14 @@ const COORDINATE_SYSTEM_GLSL_CONSTANTS = Object.keys(COORDINATE_SYSTEM)
 const PROJECTION_MODE_GLSL_CONSTANTS = Object.keys(PROJECTION_MODE)
   .map(key => `const int PROJECTION_MODE_${key} = ${PROJECTION_MODE[key]};`)
   .join('');
+const UNIT_GLSL_CONSTANTS = Object.keys(UNIT)
+  .map(key => `const int UNIT_${key.toUpperCase()} = ${UNIT[key]};`)
+  .join('');
 
 export default `\
 ${COORDINATE_SYSTEM_GLSL_CONSTANTS}
 ${PROJECTION_MODE_GLSL_CONSTANTS}
+${UNIT_GLSL_CONSTANTS}
 
 uniform int project_uCoordinateSystem;
 uniform int project_uProjectionMode;
@@ -47,6 +51,8 @@ uniform float project_uDevicePixelRatio;
 uniform float project_uFocalDistance;
 uniform vec3 project_uCameraPosition;
 uniform vec3 project_uCoordinateOrigin;
+uniform vec3 project_uCommonOrigin;
+uniform bool project_uPseudoMeters;
 
 const float TILE_SIZE = 512.0;
 const float PI = 3.1415926536;
@@ -55,20 +61,48 @@ const vec3 ZERO_64_LOW = vec3(0.0);
 const float EARTH_RADIUS = 6370972.0; // meters
 const float GLOBE_RADIUS = 256.0;
 
+// returns an adjustment factor for uCommonUnitsPerMeter
+float project_size() {
+  if (project_uProjectionMode == PROJECTION_MODE_WEB_MERCATOR &&
+    project_uCoordinateSystem == COORDINATE_SYSTEM_LNGLAT &&
+    project_uPseudoMeters == false) {
+
+    // uCommonUnitsPerMeter in low-zoom Web Mercator is non-linear
+    // Adjust by 1 / cos(latitude)
+    // If geometry.position (vertex in common space) is populated, use it
+    // Otherwise use geometry.worldPosition (anchor in world space)
+    
+    if (geometry.position.w == 0.0) {
+      float y = clamp(geometry.worldPosition.y, -89.9, 89.9);
+      return 1.0 / cos(radians(y));
+    }
+
+    // latitude from common y: 2.0 * (atan(exp(y / TILE_SIZE * 2.0 * PI - PI)) - PI / 4.0)
+    // Taylor series of 1 / cos(latitude)
+    // Max error < 0.003
+  
+    float y = geometry.position.y / TILE_SIZE * 2.0 - 1.0;
+    float y2 = y * y;
+    float y4 = y2 * y2;
+    float y6 = y4 * y2;
+    return 1.0 + 4.9348 * y2 + 4.0587 * y4 + 1.5642 * y6;
+  }
+  return 1.0;
+}
 //
 // Scaling offsets - scales meters to "world distance"
 // Note the scalar version of project_size is for scaling the z component only
 //
 float project_size(float meters) {
-  return meters * project_uCommonUnitsPerMeter.z;
+  return meters * project_uCommonUnitsPerMeter.z * project_size();
 }
 
 vec2 project_size(vec2 meters) {
-  return meters * project_uCommonUnitsPerMeter.xy;
+  return meters * project_uCommonUnitsPerMeter.xy * project_size();
 }
 
 vec3 project_size(vec3 meters) {
-  return meters * project_uCommonUnitsPerMeter;
+  return meters * project_uCommonUnitsPerMeter * project_size();
 }
 
 vec4 project_size(vec4 meters) {
@@ -87,9 +121,6 @@ vec3 project_normal(vec3 vector) {
 
 vec4 project_offset_(vec4 offset) {
   float dy = offset.y;
-  if (project_uCoordinateSystem == COORDINATE_SYSTEM_LNGLAT) {
-    dy = clamp(dy, -1., 1.);
-  }
   vec3 commonUnitsPerWorldUnit = project_uCommonUnitsPerWorldUnit + project_uCommonUnitsPerWorldUnit2 * dy;
   return vec4(offset.xyz * commonUnitsPerWorldUnit, offset.w);
 }
@@ -106,7 +137,7 @@ vec2 project_mercator_(vec2 lnglat) {
   return vec2(
     radians(x) + PI,
     PI + log(tan_fp32(PI * 0.25 + radians(y) * 0.5))
-  );
+  ) * WORLD_SCALE;
 }
 
 vec3 project_globe_(vec3 lnglatz) {
@@ -132,7 +163,7 @@ vec4 project_position(vec4 position, vec3 position64Low) {
   if (project_uProjectionMode == PROJECTION_MODE_WEB_MERCATOR) {
     if (project_uCoordinateSystem == COORDINATE_SYSTEM_LNGLAT) {
       return vec4(
-        project_mercator_(position_world.xy) * WORLD_SCALE,
+        project_mercator_(position_world.xy),
         project_size(position_world.z),
         position_world.w
       );
@@ -147,6 +178,19 @@ vec4 project_position(vec4 position, vec3 position64Low) {
         project_globe_(position_world.xyz),
         position_world.w
       );
+    }
+  }
+  if (project_uProjectionMode == PROJECTION_MODE_WEB_MERCATOR_AUTO_OFFSET) {
+    if (project_uCoordinateSystem == COORDINATE_SYSTEM_LNGLAT) {
+      if (abs(position_world.y - project_uCoordinateOrigin.y) > 0.25) {
+        // Too far from the projection center for offset mode to be accurate
+        // Only use high parts
+        return vec4(
+          project_mercator_(position_world.xy) - project_uCommonOrigin.xy,
+          project_size(position_world.z),
+          position_world.w
+        );
+      }
     }
   }
   if (project_uProjectionMode == PROJECTION_MODE_IDENTITY ||
@@ -200,6 +244,12 @@ vec2 project_pixel_size_to_clipspace(vec2 pixels) {
 
 float project_size_to_pixel(float meters) {
   return project_size(meters) * project_uScale;
+}
+float project_size_to_pixel(float size, int unit) {
+  if (unit == UNIT_METERS) return project_size_to_pixel(size);
+  if (unit == UNIT_COMMON) return size * project_uScale;
+  // UNIT_PIXELS
+  return size;
 }
 float project_pixel_size(float pixels) {
   return pixels / project_uScale;