+++ modules/geo-layers/src/mvt-layer/mvt-layer.js
@@ -18,7 +18,7 @@ const defaultProps = {
   uniqueIdProperty: {type: 'string', value: ''},
   highlightedFeatureId: null,
   loaders: [MVTWorkerLoader],
-  binary: false
+  binary: true
 };
 
 export default class MVTLayer extends TileLayer {
@@ -198,7 +198,7 @@ export default class MVTLayer extends TileLayer {
   getHighlightedObjectIndex(tile) {
     const {hoveredFeatureId, hoveredFeatureLayerName} = this.state;
     const {uniqueIdProperty, highlightedFeatureId, binary} = this.props;
-    const {data} = tile;
+    const data = tile.content;
 
     const isHighlighted = isFeatureIdDefined(highlightedFeatureId);
     const isFeatureIdPresent = isFeatureIdDefined(hoveredFeatureId) || isHighlighted;