+++ modules/layers/src/arc-layer/arc-layer.js
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import {Layer, project32, picking} from '@deck.gl/core';
+import {Layer, project32, picking, UNIT} from '@deck.gl/core';
 
 import GL from '@luma.gl/constants';
 import {Model, Geometry} from '@luma.gl/core';
@@ -124,7 +124,6 @@ export default class ArcLayer extends Layer {
   }
 
   draw({uniforms}) {
-    const {viewport} = this.context;
     const {
       widthUnits,
       widthScale,
@@ -134,13 +133,12 @@ export default class ArcLayer extends Layer {
       wrapLongitude
     } = this.props;
 
-    const widthMultiplier = widthUnits === 'pixels' ? viewport.metersPerPixel : 1;
-
     this.state.model
       .setUniforms(uniforms)
       .setUniforms({
         greatCircle,
-        widthScale: widthScale * widthMultiplier,
+        widthUnits: UNIT[widthUnits],
+        widthScale,
         widthMinPixels,
         widthMaxPixels,
         useShortestPath: wrapLongitude