+++ modules/layers/src/column-layer/column-layer-vertex.glsl.js
@@ -47,6 +47,8 @@ uniform float edgeDistance;
 uniform float widthScale;
 uniform float widthMinPixels;
 uniform float widthMaxPixels;
+uniform int radiusUnits;
+uniform int widthUnits;
 
 // Result
 varying vec4 vColor;
@@ -67,7 +69,8 @@ void main(void) {
   if (extruded) {
     elevation = instanceElevations * (positions.z + 1.0) / 2.0 * elevationScale;
   } else if (isStroke) {
-    float widthPixels = clamp(project_size_to_pixel(instanceStrokeWidths * widthScale),
+    float widthPixels = clamp(
+      project_size_to_pixel(instanceStrokeWidths * widthScale, widthUnits),
       widthMinPixels, widthMaxPixels) / 2.0;
     strokeOffsetRatio -= sign(positions.z) * project_pixel_size(widthPixels) / project_size(edgeDistance * coverage * radius);
   }
@@ -82,7 +85,11 @@ void main(void) {
   // project center of column
   vec3 centroidPosition = vec3(instancePositions.xy, instancePositions.z + elevation);
   vec3 centroidPosition64Low = instancePositions64Low;
-  vec3 pos = vec3(project_size(rotationMatrix * positions.xy * strokeOffsetRatio + offset) * dotRadius, 0.);
+  vec2 offset = (rotationMatrix * positions.xy * strokeOffsetRatio + offset) * dotRadius;
+  if (radiusUnits == UNIT_METERS) {
+    offset = project_size(offset);
+  }
+  vec3 pos = vec3(offset, 0.);
   DECKGL_FILTER_SIZE(pos, geometry);
 
   gl_Position = project_position_to_clipspace(centroidPosition, centroidPosition64Low, pos, geometry.position);