+++ modules/layers/src/scatterplot-layer/scatterplot-layer-vertex.glsl.js
@@ -41,6 +41,8 @@ uniform float lineWidthMaxPixels;
 uniform float stroked;
 uniform bool filled;
 uniform bool billboard;
+uniform int radiusUnits;
+uniform int lineWidthUnits;
 
 varying vec4 vFillColor;
 varying vec4 vLineColor;
@@ -53,13 +55,13 @@ void main(void) {
 
   // Multiply out radius and clamp to limits
   outerRadiusPixels = clamp(
-    project_size_to_pixel(radiusScale * instanceRadius),
+    project_size_to_pixel(radiusScale * instanceRadius, radiusUnits),
     radiusMinPixels, radiusMaxPixels
   );
   
   // Multiply out line width and clamp to limits
   float lineWidthPixels = clamp(
-    project_size_to_pixel(lineWidthScale * instanceLineWidths),
+    project_size_to_pixel(lineWidthScale * instanceLineWidths, lineWidthUnits),
     lineWidthMinPixels, lineWidthMaxPixels
   );