+++ modules/layers/src/text-layer/text-background-layer/text-background-layer.js
@@ -1,4 +1,4 @@
-import {Layer, project32, picking} from '@deck.gl/core';
+import {Layer, project32, picking, UNIT} from '@deck.gl/core';
 import GL from '@luma.gl/constants';
 import {Model, Geometry} from '@luma.gl/core';
 
@@ -94,7 +94,6 @@ export default class TextBackgroundLayer extends Layer {
   }
 
   draw({uniforms}) {
-    const {viewport} = this.context;
     const {
       billboard,
       sizeScale,
@@ -105,7 +104,6 @@ export default class TextBackgroundLayer extends Layer {
     } = this.props;
     let {padding} = this.props;
 
-    const sizeScaleMultiplier = sizeUnits === 'pixels' ? viewport.metersPerPixel : 1;
     if (padding.length < 4) {
       padding = [padding[0], padding[1], padding[0], padding[1]];
     }
@@ -116,7 +114,8 @@ export default class TextBackgroundLayer extends Layer {
         billboard,
         stroked: Boolean(getLineWidth),
         padding,
-        sizeScale: sizeScale * sizeScaleMultiplier,
+        sizeUnits: UNIT[sizeUnits],
+        sizeScale,
         sizeMinPixels,
         sizeMaxPixels
       })