+++ test/modules/geo-layers/tile-layer/tile-layer.spec.js
@@ -24,6 +24,9 @@ import {ScatterplotLayer} from '@deck.gl/layers';
 import {generateLayerTests, testLayerAsync, testLayer} from '@deck.gl/test-utils';
 import {TileLayer} from '@deck.gl/geo-layers';
 
+const DUMMY_DATA =
+  'https://raw.githubusercontent.com/visgl/deck.gl-data/master/test-data/geojson-point.json';
+
 test('TileLayer', async t => {
   const testCases = generateLayerTests({
     Layer: TileLayer,
@@ -69,7 +72,7 @@ test('TileLayer', async t => {
   const testCases = [
     {
       props: {
-        data: 'http://echo.jsontest.com/type/Point'
+        data: DUMMY_DATA
       },
       onBeforeUpdate: () => {
         t.comment('Default getTileData');
@@ -93,7 +96,7 @@ test('TileLayer', async t => {
       },
       onAfterUpdate: ({layer, subLayers}) => {
         if (!layer.isLoaded) {
-          t.ok(subLayers.length < 2);
+          t.is(subLayers.length, 2, 'Cached layers are used while loading');
         } else {
           t.is(subLayers.length, 2, 'Rendered sublayers');
           t.is(getTileDataCalled, 2, 'Fetched tile data');
@@ -160,7 +163,7 @@ test('TileLayer', async t => {
       onAfterUpdate: ({layer, subLayers}) => {
         if (layer.isLoaded) {
           t.is(getTileDataCalled, 6, 'Refetched tile data');
-          t.is(subLayers.length, 4, 'Invalidated cached sublayers with prop change');
+          t.is(subLayers.length, 8, 'Cached content is used to render sub layers');
         }
       }
     }
@@ -188,7 +191,7 @@ test('TileLayer#MapView:repeat', async t => {
   const testCases = [
     {
       props: {
-        data: 'http://echo.jsontest.com/key/value',
+        data: DUMMY_DATA,
         renderSubLayers
       },
       onAfterUpdate: ({layer, subLayers}) => {
@@ -258,7 +261,7 @@ test('TileLayer#AbortRequestsOnNewLayer', async t => {
     zoom: 1,
     repeat: true
   });
-  let tileset;
+  let tiles;
 
   const testCases = [
     {
@@ -266,7 +269,7 @@ test('TileLayer#AbortRequestsOnNewLayer', async t => {
         getTileData: () => sleep(10)
       },
       onAfterUpdate: ({layer}) => {
-        tileset = layer.state.tileset;
+        tiles = layer.state.tileset._tiles;
       }
     },
     {
@@ -275,7 +278,7 @@ test('TileLayer#AbortRequestsOnNewLayer', async t => {
       },
       onAfterUpdate: () => {
         t.is(
-          tileset._tiles.map(tile => tile._isCancelled).length,
+          tiles.filter(tile => tile._isCancelled).length,
           4,
           'all tiles from discarded layer should be cancelled'
         );